+++ exonum/src/node/consensus.rs
@@ -63,7 +63,7 @@ impl NodeHandler {
             let round = msg.round();
             self.state.add_queued(msg);
             trace!("Trying to reach actual round.");
-            if let Some(r) = self.state.get_actual_round(validator, round) {
+            if let Some(r) = self.state.update_validator_round(validator, round) {
                 trace!("Scheduling jump to round.");
                 let height = self.state.height();
                 self.execute_later(InternalRequest::JumpToRound(height, r));
@@ -516,9 +516,6 @@ impl NodeHandler {
         self.broadcast_status();
         self.add_status_timeout();
 
-        // Adjust propose timeout after accepting a new block.
-        self.state.adjust_timeout(&*self.blockchain.snapshot());
-
         // Add timeout for first round
         self.add_round_timeout();
         // Send propose we is leader
@@ -640,6 +637,7 @@ impl NodeHandler {
 
         info!("Jump to a new round = {}", round);
         self.state.jump_round(round);
+        self.add_round_timeout();
         self.process_new_round();
     }