+++ exonum/src/node/mod.rs
@@ -21,7 +21,6 @@ pub use self::state::{RequestData, State, ValidatorState};
 pub use self::whitelist::Whitelist;
 
 pub mod state; // TODO: temporary solution to get access to WAIT constants (ECR-167)
-pub mod timeout_adjuster;
 
 use failure;
 use toml::Value;
@@ -153,10 +152,15 @@ pub struct NodeApiConfig {
     /// Listen address for private api endpoints.
     pub private_api_address: Option<SocketAddr>,
     /// Cross-origin resource sharing ([CORS][cors]) options for responses returned
-    /// by API handlers.
+    /// by public API handlers.
     ///
     /// [cors]: https://developer.mozilla.org/en-US/docs/Web/HTTP/CORS
-    pub allow_origin: Option<AllowOrigin>,
+    pub public_allow_origin: Option<AllowOrigin>,
+    /// Cross-origin resource sharing ([CORS][cors]) options for responses returned
+    /// by private API handlers.
+    ///
+    /// [cors]: https://developer.mozilla.org/en-US/docs/Web/HTTP/CORS
+    pub private_allow_origin: Option<AllowOrigin>,
 }
 
 impl Default for NodeApiConfig {
@@ -166,7 +170,8 @@ impl Default for NodeApiConfig {
             enable_blockchain_explorer: true,
             public_api_address: None,
             private_api_address: None,
-            allow_origin: None,
+            public_allow_origin: None,
+            private_allow_origin: None,
         }
     }
 }
@@ -409,7 +414,7 @@ impl NodeHandler {
 
         let mut whitelist = config.listener.whitelist;
         whitelist.set_validators(stored.validator_keys.iter().map(|x| x.consensus_key));
-        let mut state = State::new(
+        let state = State::new(
             validator_id,
             config.listener.consensus_public_key,
             config.listener.consensus_secret_key,
@@ -424,9 +429,6 @@ impl NodeHandler {
             last_height,
             system_state.current_time(),
         );
-
-        // Adjust propose timeout for the first time.
-        state.adjust_timeout(&*snapshot);
 
         NodeHandler {
             blockchain,
@@ -462,6 +464,21 @@ impl NodeHandler {
     /// Returns value of the `txs_block_limit` field from the current `ConsensusConfig`.
     pub fn txs_block_limit(&self) -> u32 {
         self.state().consensus_config().txs_block_limit
+    }
+
+    /// Returns value of the minimal propose timeout.
+    pub fn min_propose_timeout(&self) -> Milliseconds {
+        self.state().consensus_config().min_propose_timeout
+    }
+
+    /// Returns value of the maximal propose timeout.
+    pub fn max_propose_timeout(&self) -> Milliseconds {
+        self.state().consensus_config().max_propose_timeout
+    }
+
+    /// Returns threshold starting from which the minimal propose timeout value is used.
+    pub fn propose_timeout_threshold(&self) -> u32 {
+        self.state().consensus_config().propose_timeout_threshold
     }
 
     /// Returns `State` of the node.
@@ -584,9 +601,17 @@ impl NodeHandler {
 
     /// Adds `NodeTimeout::Propose` timeout to the channel.
     pub fn add_propose_timeout(&mut self) {
-        let adjusted_timeout = self.state.propose_timeout();
-        let time =
-            self.round_start_time(self.state.round()) + Duration::from_millis(adjusted_timeout);
+        let snapshot = self.blockchain.snapshot();
+        let timeout = if Schema::new(&snapshot).transactions_pool_len()
+            >= self.propose_timeout_threshold() as usize
+        {
+            self.min_propose_timeout()
+        } else {
+            self.max_propose_timeout()
+        };
+        self.state.set_propose_timeout(timeout);
+
+        let time = self.round_start_time(self.state.round()) + Duration::from_millis(timeout);
 
         trace!(
             "ADD PROPOSE TIMEOUT: time={:?}, height={}, round={}",
@@ -872,8 +897,12 @@ impl Node {
         // Start private api.
         if let Some(listen_address) = self.api_options.private_api_address {
             let api_sender = self.channel();
-            let handler =
-                create_private_api_handler(blockchain.clone(), api_state.clone(), api_sender);
+            let handler = create_private_api_handler(
+                blockchain.clone(),
+                api_state.clone(),
+                api_sender,
+                &self.api_options,
+            );
             let listener = Iron::new(handler).http(listen_address).unwrap();
             api_handlers.push(listener);
 
@@ -971,7 +1000,7 @@ pub fn create_public_api_handler(
     mount.mount("api/system", router);
 
     let mut chain = Chain::new(mount);
-    if let Some(ref allow_origin) = config.allow_origin {
+    if let Some(ref allow_origin) = config.public_allow_origin {
         chain.link_around(CorsMiddleware::from(allow_origin.clone()));
     }
     chain
@@ -983,6 +1012,7 @@ pub fn create_private_api_handler(
     blockchain: Blockchain,
     shared_api_state: SharedNodeState,
     api_sender: ApiSender,
+    config: &NodeApiConfig,
 ) -> Chain {
     let mut mount = Mount::new();
     mount.mount("api/services", blockchain.mount_private_api());
@@ -993,7 +1023,11 @@ pub fn create_private_api_handler(
     system_api.wire(&mut router);
     mount.mount("api/system", router);
 
-    Chain::new(mount)
+    let mut chain = Chain::new(mount);
+    if let Some(ref allow_origin) = config.private_allow_origin {
+        chain.link_around(CorsMiddleware::from(allow_origin.clone()));
+    }
+    chain
 }
 
 #[cfg(test)]