+++ CHANGELOG.md
@@ -5,9 +5,30 @@ The project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
 
 ## [Unreleased]
 
+### Breaking changes
+
+#### exonum
+
+- `TimeoutAdjusterConfig` has been removed along with different timeout
+  adjusters. Current behavior is similar to the `Dynamic` timeout adjuster and
+  can be modified through `min_propose_timeout`, `max_propose_timeout` and
+  `propose_timeout_threshold` fields in the `ConsensusConfig`. (#643)
+
+  Migration path:
+
+  - `Constant` timeout adjuster can be emulated by setting equal
+  `min_propose_timeout` and `max_propose_timeout` values.
+  - For `Dynamic` timeout adjuster simply move `min`, `max` and `threshold`
+    values into `min_propose_timeout`, `max_propose_timeout` and
+    `propose_timeout_threshold` correspondingly.
+  - There is no possibility to emulate `MovingAverage` now, so `Dynamic` should
+    be used as the closest alternative.
+
 ### New features
 
 #### exonum
+
+- Private API now support CORS. (#675)
 
 - The `--allow-origin` parameter has been added to the `finalize` command.
 
@@ -16,8 +37,13 @@ The project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
 - `Field`, `CryptoHash`, `StorageValue` and `ExonumJson` traits have been
   implemented for `chrono::Duration` structure. (#653)
 
+- `execute` method has been added in `Service` trait. (#667)
+
 - `Field`, `CryptoHash`, `StorageKey`, `StorageValue` and `ExonumJson` traits
   have been implemented for `rust_decimal::Decimal`. (#671)
+
+- Maintenance CLI command for node management has been added. Currently the only
+  supported command is `clear-cache` which clears node message cache. (#676)
 
 #### exonum-timestamping
 
@@ -31,6 +57,10 @@ The project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
   `NodeHandler::handle_incoming_tx` method. (#642)
 
 - Fixed bug with shutdown requests handling. (#666)
+
+- Fixed deserialization of the `MapProof` data structure. (#674)
+
+- Fixed a bug which prevented the node from reaching the actual round. (#680 #681)
 
 #### exonum-cryptocurrency-advanced